improved equalizer precision
git-svn-id: svn://svn.code.sf.net/p/qmmp-dev/code/trunk/qmmp@689 90c681e8-e032-0410-971d-27865f9a5e38

+++ src/plugins/Input/cue/decoder_cue.cpp
@@ -220,7 +220,7 @@ void DecoderCUE::pause()
     }
 }
 
-void DecoderCUE::setEQ(int bands[10], int preamp)
+void DecoderCUE::setEQ(double bands[10], double preamp)
 {
     for (int i = 0; i < 10; ++i)
         m_bands2[i] = bands[i];
